+++ impacted-test-engine/src/test/java/com/teamscale/test_impacted/engine/InternalImpactedTestEngineTest.java
@@ -74,6 +74,8 @@ class InternalImpactedTestEngineTest {
 		when(testEngine.getId()).thenReturn("junit-jupiter");
 		when(testwiseCoverageAgentApi.testStarted(anyString())).thenReturn(mock(Call.class));
 		when(testwiseCoverageAgentApi.testFinished(anyString())).thenReturn(mock(Call.class));
+		when(testwiseCoverageAgentApi.testFinished(anyString(), any())).thenReturn(mock(Call.class));
+		when(testwiseCoverageAgentApi.testRunFinished()).thenReturn(mock(Call.class));
 	}
 
 	@SafeVarargs